refactor: extract room domain helpers
